Moved expect rule code to separate file and some final polishing.
